Fix: Make RPC compatible with 64-bit architectures
Original code used unsigned as type for arguments of RPC-related system
calls and the system call machinery. That's OK for 32-bit machines as
these are ILP32.

With LP64 platforms, unsigned remains 32-bits long, which is shorter
than 64-bits - the size of pointer. Thus data becomes damaged if
communicated via Syscall API and is larger than 32 bits.

Thus the type used for system call function prototype and especially
RPC-related functions was selected as unsigned long, which should cover
all possible scenarios even on 64-bit machines while not breaking 32-bit
builds.

+++ src/lib/rpc.c
@@ -4,7 +4,7 @@
 #include <cmrx/assert.h>
 #include <arch/sysenter.h>
 
-__SYSCALL int _rpc_call(unsigned arg0, unsigned arg1, unsigned arg2, unsigned arg3, void * service, unsigned method, unsigned canary)
+__SYSCALL int _rpc_call(unsigned long arg0, unsigned long arg1, unsigned long arg2, unsigned long arg3, void * service, unsigned method, unsigned canary)
 {
     (void) arg0;
     (void) arg1;
